Update Tour
Added a PATCH route to update a specific tour by id. The route checks if the id from the request parameters is valid by comparing it to the length of the tours array. If the id is invalid, a 404 response with an error message is sent. Otherwise, a placeholder success response is returned, indicating the tour has been updated (actual update logic will be implemented later).