Delete Tour
Added a DELETE route to remove a specific tour by id. The route checks if the id from the request parameters is valid by comparing it to the length of the tours array. If the id is invalid, a 404 response with an error message is sent. If valid, a 204 response is returned with a null data value to indicate the tour has been successfully deleted (actual deletion logic to be implemented later).

*  Also add Comments identifying routes